requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
de_core_news_md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.8.0/de_core_news_md-3.8.0-py3-none-any.whl
Django==5.1.4
dnspython==2.7.0
elastic-transport==8.17.0
elasticsearch==8.17.1
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
Flask==3.1.0
funcy==2.0
gensim==4.3.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
preshed==3.0.9
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyLDAvis==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
starlette==0.41.3
thinc==8.3.4
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.3
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
